ml/requirements.txt
joblib==0.16.0
numpy==1.19.0
Pillow==7.2.0
scikit-learn==0.23.1
scipy==1.5.1
threadpoolctl==2.1.0
